# `array_chunks` (performance feature), `alloc` improvements, code modularisation.