--- a/src/alphabet.rs
+++ b/src/alphabet.rs
@@ -1,11 +1,11 @@
 pub const TABLE: [u8; 45] = *b"0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZ $%*+-./:";
-pub fn decode(v: u8) -> Option<usize> {
-    TABLE.iter().position(|&r| r == v)
+pub fn decode(v: u8) -> Option<u8> {
+    TABLE.iter().position(|&r| r == v).map(|i| i as _)
 }
 
-pub fn encode(n: usize) -> Option<u8> {
-    TABLE.get(n).map(|&c| c)
+pub fn encode(n: u8) -> Option<u8> {
+    TABLE.get(n as usize).map(|&c| c)
 }
 
-pub const SIZE: usize = 45;
-pub const SIZE_SIZE: usize = SIZE * SIZE;
+pub const SIZE: u32 = 45;
+pub const SIZE_SIZE: u32 = SIZE * SIZE;