--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1,4 +1,5 @@
 #![cfg_attr(feature = "array_chunks", feature(array_chunks))]
+#![cfg_attr(test, feature(test))]
 //! Encoder/decoder for base45 that is fully compatible with
 //! [`draft-faltstrom-base45-07.txt`](https://www.ietf.org/archive/id/draft-faltstrom-base45-07.txt)
 //!
@@ -22,5 +23,7 @@ mod encode;
 pub use decode::{decode, DecodeError};
 pub use encode::{encode, encode_from_buffer};
 
+#[cfg(test)]
+extern crate test;
 #[cfg(test)]
 mod tests;