- Implements the algorithm using `array_chunks`/`chunks_exact` for branchless performance improvements, and potentially reduced code size for constrained devices.
- Improvements to initial allocation for `String` and `Vec` to reduce allocations during runtime.
- Splits runtime code into `decode`, `encode`, and `test` modules.