@ZaneHannanAU 2021-07-09
Old:

```
test result: ok. 0 passed; 0 failed; 15 ignored; 0 measured; 0 filtered out; finished in 0.00s

encode long string      time:   [6.2338 us 6.2485 us 6.2658 us]                                
                        change: [+766.60% +772.21% +778.09%] (p = 0.00 < 0.05)
                        Performance has regressed.

encode long string from buffer                                                                             
                        time:   [6.3149 us 6.3315 us 6.3496 us]
                        change: [+737.03% +740.47% +743.92%] (p = 0.00 < 0.05)
                        Performance has regressed.

decode long string      time:   [1.7251 us 1.7291 us 1.7335 us]                                
                        change: [+49.848% +50.557% +51.242%] (p = 0.00 < 0.05)
                        Performance has regressed.

Switched to branch 'array_chunks'
08:03:58 $ cargo criterion
#...
test tests::bench_decode_quick_brown_fox ... bench:     131,063 ns/iter (+/- 819)
test tests::bench_encode_quick_brown_fox ... bench:      66,336 ns/iter (+/- 134)
test tests::bench_encode_random_0x10     ... bench:         246 ns/iter (+/- 1)
test tests::bench_encode_random_0x100    ... bench:       2,596 ns/iter (+/- 18)
test tests::bench_encode_random_0x1000   ... bench:      38,448 ns/iter (+/- 890)
test tests::bench_encode_random_0x10000  ... bench:     621,165 ns/iter (+/- 5,123)

test result: ok. 0 passed; 0 failed; 17 ignored; 6 measured; 0 filtered out; finished in 4.21s

encode long string      time:   [720.96 ns 722.17 ns 723.59 ns]                                
                        change: [-88.622% -88.543% -88.468%] (p = 0.00 < 0.05)
                        Performance has improved.

encode long string from buffer                                                                             
                        time:   [749.84 ns 751.25 ns 752.83 ns]
                        change: [-88.157% -88.107% -88.059%] (p = 0.00 < 0.05)
                        Performance has improved.

decode long string      time:   [1.1376 us 1.1402 us 1.1429 us]                                
                        change: [-34.342% -34.040% -33.743%] (p = 0.00 < 0.05)
                        Performance has improved.

08:05:14 $
08:07:01 $ cargo criterion --features array_chunks
#...
test tests::bench_decode_quick_brown_fox ... bench:     135,537 ns/iter (+/- 8,804)
test tests::bench_encode_quick_brown_fox ... bench:      66,740 ns/iter (+/- 292)
test tests::bench_encode_random_0x10     ... bench:         250 ns/iter (+/- 5)
test tests::bench_encode_random_0x100    ... bench:       2,569 ns/iter (+/- 27)
test tests::bench_encode_random_0x1000   ... bench:      38,377 ns/iter (+/- 698)
test tests::bench_encode_random_0x10000  ... bench:     629,568 ns/iter (+/- 38,156)

test result: ok. 0 passed; 0 failed; 17 ignored; 6 measured; 0 filtered out; finished in 2.63s

encode long string      time:   [723.99 ns 726.25 ns 728.91 ns]                                
                        change: [-0.1916% +0.2028% +0.5912%] (p = 0.32 > 0.05)
                        No change in performance detected.

encode long string from buffer                                                                             
                        time:   [744.31 ns 745.91 ns 747.66 ns]
                        change: [-1.1878% -0.7567% -0.3479%] (p = 0.00 < 0.05)
                        Change within noise threshold.

decode long string      time:   [1.1544 us 1.1571 us 1.1601 us]                                
                        change: [+1.0319% +1.4711% +1.9136%] (p = 0.00 < 0.05)
                        Performance has regressed.

08:08:20 $
```

Unsure as to how this handles. Relatively, there's not too much that needs work, I think...
```

With other